@@ -1,11 +1,11 @@
-import { PipeTransform, Pipe } from '@angular/core';
+import {Pipe, PipeTransform} from '@angular/core';
 
 @Pipe({
-  name: 'adult'
+    name: 'adult'
 })
 
 export class AdultPipe implements PipeTransform {
-  transform(value) {
-    return value.filter(x => x.age > 18);
-  }
+    transform(value) {
+        return value.filter(x => x.age > 18);
+    }
 }
