@@ -1,5 +1,5 @@
-import { Component, OnInit } from '@angular/core';
-import { authors, Author } from '../authors';
+import {Component, OnInit} from '@angular/core';
+import {Author, authors} from '../authors';
 
 @Component({
     selector: 'app-author-list',
@@ -11,7 +11,8 @@ export class AuthorListComponent implements OnInit {
     authors = authors;
     currentAuthor = authors[0];
 
-    constructor() { }
+    constructor() {
+    }
 
     ngOnInit(): void {
     }
@@ -24,7 +25,7 @@ export class AuthorListComponent implements OnInit {
         this.authors = this.authors.filter(author => {
             return author.id !== id;
         });
-        console.log({ author: this.authors[0].firstName });
+        console.log({author: this.authors[0].firstName});
 
         if (this.currentAuthor.id === id) {
             this.currentAuthor = this.authors[0];
