@@ -1,18 +1,18 @@
-import { Component } from '@angular/core';
+import {Component} from '@angular/core';
 
 let _count = 1;
 
 @Component({
-  selector: 'app-counter',
-  template: `count: {{ count }}`,
-  styles: [
-    `
+    selector: 'app-counter',
+    template: `count: {{ count }}`,
+    styles: [
+        `
       :host {
         display: block;
       }
     `,
-  ],
+    ],
 })
 export class CounterComponent {
-  count = _count++;
+    count = _count++;
 }
