@@ -1,48 +1,56 @@
-import { singleton, scoped, Lifecycle, container } from 'tsyringe';
+import {container, Lifecycle, scoped, singleton} from 'tsyringe';
 
 interface ProductModel {
-  sku: string;
-  name: string;
-  price: number;
+    sku: string;
+    name: string;
+    price: number;
 }
 
 interface CartItem {
-  product: ProductModel;
-  quantity: number;
+    product: ProductModel;
+    quantity: number;
 }
 
 @singleton()
 class CartService {
-  cartServiceRandomId = Math.random();
-  selectedProducts: CartItem[] = [];
-  calculateTotal(): number {
-    return 1;
-  }
-  addToCart(): void { }
+    cartServiceRandomId = Math.random();
+    selectedProducts: CartItem[] = [];
+
+    calculateTotal(): number {
+        return 1;
+    }
+
+    addToCart(): void {
+    }
 }
 
 @scoped(Lifecycle.ResolutionScoped)
 class ProductComponent {
-  productComponentRandomId = Math.random();
-  constructor(public cartService: CartService) { }
+    productComponentRandomId = Math.random();
+
+    constructor(public cartService: CartService) {
+    }
 }
 
 function testContainer() {
-  console.log(container.resolve(ProductComponent));
+    console.log(container.resolve(ProductComponent));
 }
 
 testContainer();
 
 class TestCartService {
-  selectedProducts: CartItem[] = [];
-  calculateTotal(): number {
-    return 1;
-  }
-  addToCart(): void { }
+    selectedProducts: CartItem[] = [];
+
+    calculateTotal(): number {
+        return 1;
+    }
+
+    addToCart(): void {
+    }
 }
 
 function setupTestContainer() {
-  container.register(CartService, { useClass: TestCartService });
+    container.register(CartService, {useClass: TestCartService});
 }
 
 setupTestContainer();
