@@ -1,8 +1,8 @@
-import { Directive, TemplateRef } from '@angular/core';
+import {Directive} from '@angular/core';
 
 @Directive({
-  selector: 'ng-template[tabContent]',
+    selector: 'ng-template[tabContent]',
 })
 export class TabContentDirective {
-  // constructor(public templateRef: TemplateRef<unknown>) { }
+    // constructor(public templateRef: TemplateRef<unknown>) { }
 }
